Reorganize LUTs and a bit of structural cleanup
